Add site url WordPress constants
It would be easier to have these constants set, as most of the sites that spin up are existing, which have set urls in the `options` table. I don't know whether PH has a variable for the local url, but `{{site_name}}` is a start. The TLD (`.dev`) is probably problematic as well as the set protocol I have (`http`). Happy to get more information or open up a dialogue.